requirements.txt
asgiref==3.6.0
astroid==2.15.5
autopep8==2.0.2
bcrypt==3.2.0
blinker==1.4
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
cryptography==41.0.1
dbus-python==1.2.18
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distro==1.7.0
Django==4.2.1
django-extensions==3.2.1
django-taggit==4.0.0
duplicity==0.8.21
fasteners==0.14.1
filelock==3.9.0
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
isort==5.12.0
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
package-name==0.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.5.0
pipenv==2023.4.29
platformdirs==3.1.1
protobuf==3.12.4
psycopg==3.1.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydocstyle==6.3.0
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.17.4
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
SecretStorage==3.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
systemd-python==234
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
urllib3==1.26.5
virtualenv==20.21.0
virtualenv-clone==0.5.7
wadllib==1.3.6
wrapt==1.15.0
xdg==5
zipp==1.0.0
